scenarios/handson_eventhubs/pyproject.toml
[tool.poetry]
name = "handson-eventhubs"
version = "0.0.1"
description = ""
authors = ["ks6088ts <ks6088ts@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.11.0"
python-dotenv = "^1.0.1"
azure-eventhub = "^5.11.6"
azure-eventhub-checkpointstoreblob-aio = "^1.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
